fix: improve landing page UX and design consistency
- Enhanced navbar scroll transition with smoother cubic-bezier animation (0.5s)
- Replaced fictional quality badges with real service metrics (12+ years experience, 15+ sectors, 24/7 support)
- Added 9th service: K3 (Keselamatan & Kesehatan Kerja) with complete metadata
- Fixed Cara Kerja Kami section alignment using flexbox for symmetrical layout
- Reduced section padding by 17-19% (40-64px) for better visual density
- Updated Newsletter color scheme to match LinkedIn blue theme (#0077B5, #005582)
- Improved overall design consistency across all components

+++ app/Http/Controllers/Mobile/DashboardController.php
@@ -199,10 +199,10 @@ class DashboardController extends Controller
      */
     private function getRunwayMetric()
     {
-        $totalCash = CashAccount::sum('current_balance');
+        $totalCash = (float) CashAccount::sum('current_balance');
         
         // Calculate monthly burn rate (last 30 days)
-        $monthlyExpenses = ProjectExpense::where('expense_date', '>=', now()->subDays(30))
+        $monthlyExpenses = (float) ProjectExpense::where('expense_date', '>=', now()->subDays(30))
             ->sum('amount');
             
         $monthlyBurn = $monthlyExpenses;
@@ -286,10 +286,10 @@ class DashboardController extends Controller
                 $query->where('is_active', true);
             })->count(),
             'teamMembers' => \App\Models\User::where('is_active', true)->count(),
-            'monthRevenue' => Invoice::whereMonth('created_at', now()->month)
+            'monthRevenue' => (float) Invoice::whereMonth('created_at', now()->month)
                 ->where('status', 'paid')
                 ->sum('total_amount'),
-            'monthExpenses' => ProjectExpense::whereMonth('expense_date', now()->month)
+            'monthExpenses' => (float) ProjectExpense::whereMonth('expense_date', now()->month)
                 ->sum('amount'),
         ];
     }
